avoid dead threads with file database
this showed up particularly in sporadic failures when running specs. the
"set message timeout and wait for message to come" spec would fail a
significant portion of the time when run under the file database, but
not always.

the root cause is that contention for the store's transaction between
the action processing threads and the API's @timer thread would cause
the @timer thread to die silently with a "nested transaction"
PStore::Error exception (see [1]). the YAML::Store used by the file
database inherits from PStore. while I haven't noticed it, as I use the
memory database, this same behavior could easily arise in non-test usage
as well.

the easiest solution is to mark the PStore as thread-safe, which imposes
a minor performance hit but allows normal lock semantics between
contending threads. it will then only raise the "nested transaction"
exception if the same thread tries to enter a legitimately nested
transaction. this is the solution employed for ruby 2.4+.

unfortunately, while older rubies allow creating a thread-safe PStore,
they do not allow a thread-safe YAML::Store. so we have to take a
more... intricate approach. I commented it heavily so hopefully it makes
sense.

[1] https://ruby-doc.org/stdlib-2.4.1/libdoc/pstore/rdoc/PStore.html#method-i-transaction)